Introduced Textual I/O as Media and Heads
Adapted the command sets and interpreters to use these. The result is that inputs and outputs are not necessarily shared amongst interpreters anymore.

Some work on abstracting heads and introducing a 2D head, needed for languages like Befunge, and media like the output picture from drawfuck.

Adjusted the tests to match the changes in the design